+++ wayland-client/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Bump bitflags to 2.0
 - Updated wayland-backend to 0.2
+- Calloop integration is now removed, to avoid tying wayland-client to it
 
 ## 0.30.2 -- 30/05/2023